Add response code, better message, and URI to Solarium errors.
Recently we have had some potential network problems regarding our
solr infrastructure. When an error occurs, the logs do not have much
detail regarding the error.

The changes here adds the URI to the error log of all failed requests
as well as response code, query request, and message. This should help
us in troubleshooting issues, not to mention we can now easily copy
and pase the URI into a browser and replicate the request
immediately.